Added seeding data, added RBAC

+++ src/routes/acronym.routes.js
@@ -2,10 +2,47 @@ const express = require("express");
 const router = express.Router();
 
 // model
+const Category = require("../models/Category");
 const Acronym = require("../models/Acronym");
+const seed = require("../seed/acronyms");
 
 // routes
+router.post("/seed", async(red, res, next) => {
+  try {
+    await Acronym.deleteMany({});
+
+    // build data
+    let data = [];
+    for(const s of seed) {
+      let categories = [];
+      for(const p of s.categories) {
+        const category = await Category.findOne({ "name": p });
+        if (category) {
+          categories.push(category._id);
+        }
+      }
+      data.push({
+        "acronym": s.acronym,
+        "definitions": s.definitions,
+        "status": s.status,
+        "categories": categories
+      });
+    }
+
+    await Acronym.insertMany(data);
+    return res.status(200).json({
+      message: "Acronyms with categories seeded successfully"
+    });
+  } catch(error) {
+    return next(error);
+  }
+});
+
 router.get("/", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view acronyms")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const all = await Acronym.find();
     return res.status(200).json(all);
@@ -15,6 +52,10 @@ router.get("/", async (req, res, next) => {
 });
 
 router.get("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view acronyms")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const single = await Acronym.findById(req.params.id);
     return res.status(200).json(single);
@@ -24,6 +65,10 @@ router.get("/:id", async (req, res, next) => {
 });
 
 router.post("/", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can add acronyms")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const created = new Acronym({
       acronym: req.body.acronym,
@@ -39,6 +84,10 @@ router.post("/", async (req, res, next) => {
 });
 
 router.put("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can edit acronyms")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const updated = await Acronym.findByIdAndUpdate(
       req.params.id,
@@ -57,6 +106,10 @@ router.put("/:id", async (req, res, next) => {
 });
 
 router.delete("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can delete acronyms")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const deleted = await Acronym.findByIdAndDelete(req.params.id);
     return res.status(200).json(deleted);
