+++ src/routes/category.routes.js
@@ -3,9 +3,26 @@ const router = express.Router();
 
 // model
 const Category = require("../models/Category");
+const seed = require("../seed/categories");
 
 // routes
+router.post("/seed", async(red, res, next) => {
+  try {
+    await Category.deleteMany({});
+    await Category.insertMany(seed);
+    return res.status(200).json({
+      message: "Categories seeded successfully"
+    });
+  } catch(error) {
+    return next(error);
+  }
+});
+
 router.get("/", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view categories")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const all = await Category.find();
     return res.status(200).json(all);
@@ -15,6 +32,10 @@ router.get("/", async (req, res, next) => {
 });
 
 router.get("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view categories")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const single = await Category.findById(req.params.id);
     return res.status(200).json(single);
@@ -24,9 +45,14 @@ router.get("/:id", async (req, res, next) => {
 });
 
 router.post("/", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can add categories")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const created = new Category({
-      name: req.body.name
+      name: req.body.name,
+      description: req.body.description,
     });
     const record = await created.save();
     return res.status(200).json(record);
@@ -36,10 +62,17 @@ router.post("/", async (req, res, next) => {
 });
 
 router.put("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can edit categories")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const updated = await Category.findByIdAndUpdate(
       req.params.id,
-      { name: req.body.name },
+      {
+        name: req.body.name,
+        description: req.body.description
+      },
       { new: true }
     );
     return res.status(200).json(updated);
@@ -49,6 +82,10 @@ router.put("/:id", async (req, res, next) => {
 });
 
 router.delete("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can delete categories")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const deleted = await Category.findByIdAndDelete(req.params.id);
     return res.status(200).json(deleted);
