+++ src/routes/user.routes.js
@@ -2,10 +2,47 @@ const express = require("express");
 const router = express.Router();
 
 // model
+const Role = require("../models/Role");
 const User = require("../models/User");
+const seed = require("../seed/users");
+const { hashPassword } = require("../utils/util");
 
 // routes
+router.post("/seed", async(red, res, next) => {
+  try {
+    await User.deleteMany({});
+
+    // build data
+    let data = [];
+    for(const s of seed) {
+      let roles = [];
+      for(const r of s.roles) {
+        const role = await Role.findOne({ "name": r });
+        if (role) {
+          roles.push(role._id);
+        }
+      }
+      data.push({
+        "name": s.name,
+        "email": s.email,
+        "password": await hashPassword(s.password),
+        "roles": roles
+      });
+    }
+
+    await User.insertMany(data);
+    return res.status(200).json({
+      message: "Users with roles seeded successfully"
+    });
+  } catch(error) {
+    return next(error);
+  }
+});
 router.get("/", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view users")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const all = await User.find();
     return res.status(200).json(all);
@@ -15,6 +52,10 @@ router.get("/", async (req, res, next) => {
 });
 
 router.get("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can view users")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const single = await User.findById(req.params.id);
     return res.status(200).json(single);
@@ -24,6 +65,10 @@ router.get("/:id", async (req, res, next) => {
 });
 
 router.put("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can edit users")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const updated = await User.findByIdAndUpdate(
       req.params.id,
@@ -41,6 +86,10 @@ router.put("/:id", async (req, res, next) => {
 });
 
 router.delete("/:id", async (req, res, next) => {
+  if(!req._user.permissions.includes("Can delete users")) {
+    return res.status(403).json("You are not authorized to access this resource");
+  }
+
   try {
     const deleted = await User.findByIdAndDelete(req.params.id);
     return res.status(200).json(deleted);
